% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chapter-08-au_map.R
\name{au_map}
\alias{au_map}
\title{Australia maps

\code{[figure8.1()]}: complete Australia map.
\code{[figure8.7()]}: Australia left-bottom subplot.
\code{[figure8.18()]}: Australia with temperature ribbon}
\description{
Australia maps

\code{[figure8.1()]}: complete Australia map.
\code{[figure8.7()]}: Australia left-bottom subplot.
\code{[figure8.18()]}: Australia with temperature ribbon
}
