\name{figures}
\alias{figure19.1}
\alias{figure19.2}
\alias{figure3.1}
\alias{figure3.2}
\alias{figure3.3}
\alias{figure3.4}
\alias{figure3.5}
\alias{figure3.6}
\alias{figure3.7}
\alias{figure3.8}
\alias{figure3.9}
\alias{figure3.10}
\alias{figure3.11}
\alias{figure3.12}
\alias{figure3.13}
\alias{figure3.14}
\alias{figure3.15}
\alias{figure3.16}
\alias{figure3.17}
\alias{figure3.18}
\alias{figure3.19}
\alias{figure3.20}
\alias{figure3.21}
\alias{figure3.22}
\alias{figure3.23}
\alias{figure3.24}
\alias{figure3.25}
\alias{figure3.26}
\alias{figure3.27}
\alias{figure6.1}
\alias{figure6.2}
\alias{figure6.3}
\alias{figure6.4}
\alias{figure6.5}
\alias{figure6.6}
\alias{figure6.7}
\alias{figure6.8}
\alias{figure6.9}
\alias{figure6.10}
\alias{figure6.11}
\alias{figure6.12}
\alias{figure6.13}
\alias{figure6.14}
\alias{figure6.15}
\alias{figure6.16}
\alias{figure6.17}
\alias{figure6.18}
\alias{figure6.19}
\alias{figure6.20}
\alias{figure6.21}
\alias{figure6.22}
\alias{figure6.23}
\alias{figure6.24}
\alias{figure6.25}
\alias{figure6.26}
\alias{figure1.2}
\alias{figure1.3}
\alias{figure1.4}
\alias{figure1.5}
\alias{figure1.6}
\alias{figure1.7}
\alias{figure1.8}
\alias{figure1.9}
\alias{figure1.10}
\alias{figure1.11}
\alias{figure1.12}
\alias{figure1.13}
\alias{figure11.1}
\alias{figure11.2}
\alias{figure11.3}
\alias{figure11.4}
\alias{figure11.5}
\alias{figure11.6}
\alias{figure11.7}
\alias{figure11.8}
\alias{figure11.9}
\alias{figure11.10}
\alias{figure11.11}
\alias{figure11.12}
\alias{figure11.13}
\alias{figure5.1}
\alias{figure5.2}
\alias{figure5.3}
\alias{figure5.4}
\alias{figure5.5}
\alias{figure5.6}
\alias{figure5.7}
\alias{figure5.8}
\alias{figure5.9}
\alias{figure5.10}
\alias{figure5.11}
\alias{figure5.12}
\alias{figure5.13}
\alias{figure5.14}
\alias{figure5.15}
\alias{figure15.1}
\alias{figure15.3}
\alias{figure15.4}
\alias{figure15.5}
\alias{figure15.6}
\alias{figure15.7}
\alias{figure15.8}
\alias{figure15.9}
\alias{figure15.10}
\alias{figure18.1}
\alias{figure18.2}
\alias{figure18.3}
\alias{figure18.4}
\alias{figure18.5}
\alias{figure18.6}
\alias{figure8.1}
\alias{figure8.3}
\alias{figure8.5}
\alias{figure8.6}
\alias{figure8.7}
\alias{figure8.13}
\alias{figure8.14}
\alias{figure8.18}
\alias{figure8.20}
\alias{figure8.22}
\alias{figure8.25}
\alias{figure8.26}
\alias{figure8.27}
\alias{figure17.1}
\alias{figure17.2}
\alias{figure17.3}
\alias{figure17.4}
\alias{figure17.5}
\alias{figure17.6}
\alias{figure17.7}
\alias{figure17.8}
\alias{figure17.9}
\alias{figure17.10}
\alias{figure17.11}
\alias{figure17.12}
\alias{figure17.13}
\alias{figure17.15}
\alias{figure17.16}
\alias{figure1.1}
\alias{figure14.1}
\alias{figure14.2}
\alias{figure14.3}
\alias{figure14.4}
\alias{figure14.5}
\alias{figure14.6}
\alias{figure14.7}
\alias{figure14.8}
\alias{figure14.9}
\alias{figure14.10}
\alias{figure14.11}
\alias{figure7.1}
\alias{figure7.2}
\alias{figure7.3}
\alias{figure7.4}
\alias{figure7.5}
\alias{figure7.6}
\alias{figure7.7}
\alias{figure7.8}
\alias{figure7.9}
\alias{figure7.10}
\alias{figure7.11}
\alias{figure7.12}
\alias{figure7.13}
\alias{figure1.14}
\alias{figure10.1}
\alias{figure10.2}
\alias{figure10.3}
\alias{figure10.4}
\alias{figure2.1}
\alias{figure2.2}
\alias{figure2.3}
\alias{figure2.4}
\alias{figure2.5}
\alias{figure2.6}
\alias{figure2.7}
\alias{figure2.8}
\alias{figure2.9}
\alias{figure4.1}
\alias{figure4.2}
\alias{figure4.3}
\alias{figure4.4}
\alias{figure4.5}
\alias{figure4.6}
\alias{figure4.7}
\alias{figure4.8}
\alias{figure4.9}
\alias{figure4.10}
\alias{figure4.11}
\alias{figure4.12}
\alias{figure4.13}
\alias{figure12.1}
\alias{figure12.2}
\alias{figure12.3}
\alias{figure12.4}
\alias{figure12.5}
\alias{figure12.6}
\alias{figure12.7}
\alias{figure16.1}
\alias{figure16.2}
\alias{figure16.3}
\alias{figure16.4}
\alias{figure16.5}
\alias{figure16.6}
\alias{figure16.7}
\alias{figure16.8}
\alias{figure16.9}
\alias{figure16.10}
\alias{figure16.11}
\alias{figure13.1}
\alias{figure13.2}
\alias{figure13.3}
\alias{figure13.4}
\alias{figure13.5}
\alias{figure13.6}
\alias{figure13.7}
\alias{figure13.8}
\title{Figures for R Graphics}
\description{Functions to produce the figures from the book R Graphics.  Function \emph{i} is produced by figure\emph{i}().}
\usage{
figure19.1()
figure19.2()
figure3.1()
figure3.2()
figure3.3()
figure3.4()
figure3.5()
figure3.6()
figure3.7()
figure3.8()
figure3.9()
figure3.10()
figure3.11()
figure3.12()
figure3.13()
figure3.14()
figure3.15()
figure3.16()
figure3.17()
figure3.18()
figure3.19()
figure3.20()
figure3.21()
figure3.22()
figure3.23()
figure3.24()
figure3.25()
figure3.26()
figure3.27()
figure6.1()
figure6.2()
figure6.3()
figure6.4()
figure6.5()
figure6.6()
figure6.7()
figure6.8()
figure6.9()
figure6.10()
figure6.11()
figure6.12()
figure6.13()
figure6.14()
figure6.15()
figure6.16()
figure6.17()
figure6.18()
figure6.19()
figure6.20()
figure6.21()
figure6.22()
figure6.23()
figure6.24()
figure6.25()
figure6.26()
figure1.2()
figure1.3()
figure1.4()
figure1.5()
figure1.6()
figure1.7()
figure1.8()
figure1.9()
figure1.10()
figure1.11()
figure1.12()
figure1.13()
figure11.1()
figure11.2()
figure11.3()
figure11.4()
figure11.5()
figure11.6()
figure11.7()
figure11.8()
figure11.9()
figure11.10()
figure11.11()
figure11.12()
figure11.13()
figure5.1()
figure5.2()
figure5.3()
figure5.4()
figure5.5()
figure5.6()
figure5.7()
figure5.8()
figure5.9()
figure5.10()
figure5.11()
figure5.12()
figure5.13()
figure5.14()
figure5.15()
figure15.1()
figure15.3()
figure15.4()
figure15.5()
figure15.6()
figure15.7()
figure15.8()
figure15.9()
figure15.10()
figure18.1()
figure18.2()
figure18.3()
figure18.4()
figure18.5()
figure18.6()
figure8.1()
figure8.3()
figure8.5()
figure8.6()
figure8.7()
figure8.13()
figure8.14()
figure8.18()
figure8.20()
figure8.22()
figure8.25()
figure8.26()
figure8.27()
figure17.1()
figure17.2()
figure17.3()
figure17.4()
figure17.5()
figure17.6()
figure17.7()
figure17.8()
figure17.9()
figure17.10()
figure17.11()
figure17.12()
figure17.13()
figure17.15()
figure17.16()
figure1.1()
figure14.1()
figure14.2()
figure14.3()
figure14.4()
figure14.5()
figure14.6()
figure14.7()
figure14.8()
figure14.9()
figure14.10()
figure14.11()
figure7.1()
figure7.2()
figure7.3()
figure7.4()
figure7.5()
figure7.6()
figure7.7()
figure7.8()
figure7.9()
figure7.10()
figure7.11()
figure7.12()
figure7.13()
figure1.14()
figure10.1()
figure10.2()
figure10.3()
figure10.4()
figure2.1()
figure2.2()
figure2.3()
figure2.4()
figure2.5()
figure2.6()
figure2.7()
figure2.8()
figure2.9()
figure4.1()
figure4.2()
figure4.3()
figure4.4()
figure4.5()
figure4.6()
figure4.7()
figure4.8()
figure4.9()
figure4.10()
figure4.11()
figure4.12()
figure4.13()
figure12.1()
figure12.2()
figure12.3()
figure12.4()
figure12.5()
figure12.6()
figure12.7()
figure16.1()
figure16.2()
figure16.3()
figure16.4()
figure16.5()
figure16.6()
figure16.7()
figure16.8()
figure16.9()
figure16.10()
figure16.11()
figure13.1()
figure13.2()
figure13.3()
figure13.4()
figure13.5()
figure13.6()
figure13.7()
figure13.8()
}
\keyword{internal}
