% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid.imageGrob.R
\name{grid.imageFun}
\alias{grid.imageFun}
\title{Draw an Image

Draws an array of nrow by ncol rectangles.}
\usage{
grid.imageFun(nrow, ncol, cols, byrow = TRUE)
}
\arguments{
\item{nrow, ncol}{Number of rows/columns in the image.}

\item{cols}{Colors for the cells of the image (will be recycled).}

\item{byrow}{Logical value indicating whether colors should be allocated to
cells across rows or down columns.}
}
\description{
Draw an Image

Draws an array of nrow by ncol rectangles.
}
