% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid.imageGrob.R
\name{grid.imageGrob}
\alias{grid.imageGrob}
\title{grid.imageGrob
Creates an imageGrob and then draws it.}
\usage{
grid.imageGrob(...)
}
\arguments{
\item{...}{Arguments to be passed to \link{imageGrob}.}
}
\description{
grid.imageGrob
Creates an imageGrob and then draws it.
}
