% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid.ozFun.R
\name{grid.ozFun}
\alias{grid.ozFun}
\title{Draw a Map of Australia

Draws a map of Australia (or part thereof).}
\usage{
grid.ozFun(ozRegion)
}
\arguments{
\item{ozRegion}{An object created using the \code{ozRegion} function from the
\pkg{oz} package.}
}
\description{
Draw a Map of Australia

Draws a map of Australia (or part thereof).
}
\examples{
\dontrun{
grid.ozFun(oz::ozRegion())
}
}
