\name{grid.utext}
\alias{grid.utext}
\alias{textCorners}
\title{Draw Underlined Text}
\description{Draws a single piece of text with a line beneath.}
\usage{
grid.utext(label, x=.5, y=.5, ..., name="utext") 
textCorners(x)
}
\arguments{
  \item{label}{A character value.}
  \item{x,y}{Numeric or unit value.}
  \item{...}{Further arguments passed to \code{grid.text}.}
  \item{name}{Character value.}
}
\keyword{dplot}
