\name{grid.utextvp}
\alias{grid.utextvp}
\alias{utextvp}
\title{Draw Underlined Text}
\description{Draws a single piece of text with a line beneath.}
\usage{
grid.utextvp(label, x=.5, y=.5, ..., name="utext") 
utextvp(label, x, y, ..., name="utextvp")
}
\arguments{
  \item{label}{A character value.}
  \item{x,y}{Numeric or unit value.}
  \item{...}{Further arguments passed to \code{grid.text}.}
  \item{name}{Character value.}
}
\keyword{dplot}
