\name{fluoro.predict}
\docType{data}
\alias{fluoro.predict}
\title{Predicted Surface of Fluorescence}
\description{
  These data give a prediction surface for fluorescence at the
  thermocline over a region off the coast of South Australia.
}
\usage{fluoro.predict}
\format{A list with elements: \code{x} containing longitude at 50 locations;
  \code{y} containing latitude at 50 locations;
  and \code{z} containing a 50 by 50 matrix of surface predictions.}
\references{
  S. McClatchie and T.M. Ward. (in press),
  \cite{Alongshore variation in upwelling
intensity in the eastern Great Australian Bight},
  Journal of Geophysical
Research.

}
\keyword{datasets}
