% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid.ozFun.R
\name{ozImage}
\alias{ozImage}
\title{Create an Image on a Map of Australia}
\usage{
ozImage(mapLong, mapLat, imageLong, imageLat, cols)
}
\arguments{
\item{mapLong, mapLat}{Longitude and latitude ranges describing the area of
Australia to create.}

\item{imageLong, imageLat}{Longitude and latitude ranges describing the area
that the image should occupy.}

\item{cols}{Colors for the image cells.}
}
\description{
Creates a map of Australia (or part thereof), plus an \link{imageGrob} positioned
relative to the map.
}
