% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid.ozFun.R
\name{ozKey}
\alias{ozKey}
\title{ozKey}
\usage{
ozKey(x, y, width, height, just, mapLong, mapLat)
}
\arguments{
\item{x, y, width, height, just}{The location and size of the key within the
current viewport.}

\item{mapLong}{Longitude and latitude ranges describing the area of
Australia to create.}

\item{mapLat}{Longitude and latitude ranges describing the area of
Australia to create.}
}
\description{
Creates a map of Australia with a rectangle bounding a specified region;
designed to provide a key for an ozImage.
}
