\name{ozTemp}
\docType{data}
\alias{ozTemp}
\title{Temperatures for Australian Cities}
\description{
  These data give average minimum and maximum monthly temperatures
  for several major cities in Australia.  The longitude and
  latitude for each city is also given.
}
\usage{
data(ozTemp)
}
\format{A data frame with elements:
  \code{city} names of cities;
  \code{min} and \code{max} average minimum and maximum monthly temperatures;
  \code{long} and \code{lat} longitudes and latitudes of cities.}
\source{
  Was originally \code{http://www.auinfo.com/sydney-climate.html}
  but that URL is no longer alive.
}
\keyword{datasets}
