% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{plot.newclass}
\alias{plot.newclass}
\title{A Traditional Graphics Function Template}
\usage{
\method{plot}{newclass}(x, y = NULL, main = "", sub = "",
  xlim = NULL, ylim = NULL, axes = TRUE, ann = par("ann"),
  col = par("col"), ...)
}
\description{
A template that provides a starting point for writing a new traditional
graphics function.
}
\details{
Type plot.newclass to see the body of this template.
}
