% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RGraphics-package.R
\docType{package}
\name{RGraphics-package}
\alias{RGraphics}
\alias{RGraphics-package}
\title{RGraphics: Data and Functions from the Book R Graphics, Second Edition}
\description{
Data and Functions from the book R Graphics, Second Edition. There is a function to produce each figure in the book, plus several functions, classes, and methods defined in Chapter 8.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://www.stat.auckland.ac.nz/~paul/RG2e/index.html}
}

}
\keyword{internal}
