% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitString.R
\name{splitString}
\alias{splitString}
\title{Split text into multiple lines}
\usage{
splitString(text)
}
\arguments{
\item{text}{The string to split.}
}
\description{
Splits a single string into multiple lines (by inserting line breaks) so that
the output will fit within the current viewport.
}
\examples{
figure8.20()

}
