% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitString.R
\name{splitTextGrob}
\alias{splitTextGrob}
\title{splitTextGrob}
\usage{
splitTextGrob(text, ...)
}
\arguments{
\item{text}{The string to split.}

\item{...}{Arguments passed to the \href{grid::grid.grob()}{grob} function.}
}
\description{
splitTextGrob
}
