\name{utextvpDynamic}
\alias{utextvpDynamic}
\title{Draw Underlined Text}
\description{Creates a grob representing underlined text.}
\usage{
utextvpDynamic(label, x=.5, y=.5, default.units="npc",
          just="centre", angle=0, name="utext") 
}
\arguments{
  \item{label}{A character value.}
  \item{x,y}{Numeric or unit value.}
  \item{default.units}{Units to use if location is not a unit.}
  \item{just}{Character vector indicating justification of text 
    relative to its location.}
  \item{angle}{Numeric angle of text (in degrees).}
  \item{name}{Character value.}
}
\keyword{dplot}
