\name{utextvpStatic}
\alias{utextvpStatic}
\alias{utextvpChildren}
\title{Draw Underlined Text}
\description{Creates a grob representing underlined text.}
\usage{
utextvpStatic(label, x=.5, y=.5, default.units="npc",
              angle=0, just="centre", name="utext") 
utextvpChildren(label, name) 
}
\arguments{
  \item{label}{A character value.}
  \item{x,y}{Numeric or unit value.}
  \item{default.units}{Units to use if location is not a unit.}
  \item{angle}{Numeric angle of text (in degrees).}
  \item{just}{Character vector indicating justification of text 
    relative to its location.}
  \item{name}{Character value.}
}
\keyword{dplot}
