\name{volcano.summitRoad}
\docType{data}
\alias{volcano.summitRoad}
\title{Roads on Auckland's Maunga Whau Volcano}
\description{
  These data give a prediction surface for fluorescence at the
  thermocline over a region off the coast of South Australia.
}
\usage{volcano.summitRoad}
\format{A data frame with three variables,
  \code{ax}, \code{ay}, and \code{az},
  that describe the (three dimensional) path of the
  road that provides access to the summit of the
  Maunga Whau volcano in Auckland, New Zealand.}
\source{Digitised from same map as the original \code{volcano} data set.}
\seealso{\code{\link{volcano.upDownRoad}},
\code{\link{volcano.accessRoad}},
\code{\link[datasets]{volcano}}}
\references{
  E. J. Searle (1981), \cite{City of Volcanoes}.
}
\keyword{datasets}
