\name{volcanoRoads}
\docType{data}
\alias{volcano.accessRoad}
\alias{volcano.upDownRoad}
\alias{volcano.summitRoad}
\title{Roads on Auckland's Maunga Whau Volcano}
\description{
  These data give a prediction surface for fluorescence at the
  thermocline over a region off the coast of South Australia.
}
\usage{
data(volcano.accessRoad)
data(volcano.upDownRoad)
data(volcano.summitRoad)
}
\format{A list with elements: \code{x} containing longitude at 50 locations;
  \code{y} containing latitude at 50 locations;
  and \code{z} containing a 50 by 50 matrix of surface predictions.}
\source{Digitised from same map as the original \code{volcano} data set.}
\seealso{\code{\link[datasets]{volcano}}}
\references{
  E. J. Searle (1981), \cite{City of Volcanoes}.
}
\keyword{datasets}
