\name{figures}
\alias{figure12.1}
\alias{figure12.2}
\alias{figure12.3}
\alias{figure12.4}
\alias{figure12.5}
\alias{figure12.6}
\alias{figure12.7}
\alias{figure3.1}
\alias{figure3.2}
\alias{figure3.3}
\alias{figure3.4}
\alias{figure3.5}
\alias{figure3.6}
\alias{figure3.7}
\alias{figure3.8}
\alias{figure3.9}
\alias{figure3.10}
\alias{figure3.11}
\alias{figure3.12}
\alias{figure3.13}
\alias{figure3.14}
\alias{figure3.15}
\alias{figure3.16}
\alias{figure3.17}
\alias{figure3.18}
\alias{figure3.19}
\alias{figure3.20}
\alias{figure3.21}
\alias{figure3.22}
\alias{figure3.23}
\alias{figure3.24}
\alias{figure3.25}
\alias{figure3.26}
\alias{figure3.27}
\alias{figure3.28}
\alias{figure3.29}
\alias{figure6.1}
\alias{figure6.2}
\alias{figure6.3}
\alias{figure6.4}
\alias{figure6.5}
\alias{figure6.6}
\alias{figure6.7}
\alias{figure6.8}
\alias{figure6.9}
\alias{figure6.10}
\alias{figure6.11}
\alias{figure6.12}
\alias{figure6.13}
\alias{figure6.14}
\alias{figure6.15}
\alias{figure6.16}
\alias{figure6.17}
\alias{figure6.18}
\alias{figure6.19}
\alias{figure6.20}
\alias{figure6.21}
\alias{figure6.22}
\alias{figure6.23}
\alias{figure6.24}
\alias{figure6.25}
\alias{figure6.26}
\alias{figure1.2}
\alias{figure1.3}
\alias{figure1.4}
\alias{figure1.5}
\alias{figure1.6}
\alias{figure1.7}
\alias{figure1.8}
\alias{figure1.9}
\alias{figure1.10}
\alias{figure1.11}
\alias{figure1.12}
\alias{figure1.13}
\alias{figure13.1}
\alias{figure13.2}
\alias{figure13.3}
\alias{figure13.4}
\alias{figure13.5}
\alias{figure13.6}
\alias{figure13.7}
\alias{figure13.8}
\alias{figure13.9}
\alias{figure13.10}
\alias{figure13.11}
\alias{figure13.12}
\alias{figure13.13}
\alias{figure13.14}
\alias{figure5.1}
\alias{figure5.2}
\alias{figure5.3}
\alias{figure5.4}
\alias{figure5.5}
\alias{figure5.6}
\alias{figure5.7}
\alias{figure5.8}
\alias{figure5.9}
\alias{figure5.10}
\alias{figure5.11}
\alias{figure5.12}
\alias{figure5.13}
\alias{figure5.14}
\alias{figure5.15}
\alias{figure5.16}
\alias{figure11.1}
\alias{figure11.2}
\alias{figure11.3}
\alias{figure11.4}
\alias{figure11.5}
\alias{figure11.6}
\alias{figure11.7}
\alias{figure11.8}
\alias{figure11.9}
\alias{figure11.10}
\alias{figure11.11}
\alias{figure11.12}
\alias{figure11.13}
\alias{figure11.14}
\alias{figure8.1}
\alias{figure8.3}
\alias{figure8.4}
\alias{figure8.5}
\alias{figure8.7}
\alias{figure8.8}
\alias{figure8.9}
\alias{figure8.11}
\alias{figure8.13}
\alias{figure8.14}
\alias{figure8.16}
\alias{figure8.17}
\alias{figure8.18}
\alias{figure8.20}
\alias{figure8.21}
\alias{figure8.22}
\alias{figure8.23}
\alias{figure8.25}
\alias{figure8.26}
\alias{figure8.28}
\alias{figure8.30}
\alias{figure8.32}
\alias{figure1.1}
\alias{figure7.1}
\alias{figure7.2}
\alias{figure7.3}
\alias{figure7.4}
\alias{figure7.5}
\alias{figure7.6}
\alias{figure7.7}
\alias{figure7.8}
\alias{figure7.9}
\alias{figure7.10}
\alias{figure7.11}
\alias{figure7.12}
\alias{figure7.13}
\alias{figure7.14}
\alias{figure7.15}
\alias{figure7.16}
\alias{figure1.14}
\alias{figure10.1}
\alias{figure10.2}
\alias{figure10.3}
\alias{figure10.4}
\alias{figure10.5}
\alias{figure10.6}
\alias{figure10.7}
\alias{figure10.8}
\alias{figure10.9}
\alias{figure10.10}
\alias{figure10.11}
\alias{figure10.12}
\alias{figure10.13}
\alias{figure2.1}
\alias{figure2.2}
\alias{figure2.3}
\alias{figure2.4}
\alias{figure2.5}
\alias{figure2.6}
\alias{figure2.7}
\alias{figure2.8}
\alias{figure2.9}
\alias{figure2.10}
\alias{figure4.1}
\alias{figure4.2}
\alias{figure4.3}
\alias{figure4.4}
\alias{figure4.5}
\alias{figure4.6}
\alias{figure4.7}
\alias{figure4.8}
\alias{figure4.9}
\alias{figure4.10}
\alias{figure4.11}
\alias{figure4.12}
\title{Figures for R Graphics}
\description{Functions to produce the figures from the book R Graphics.  Function \emph{i} is produced by figure\emph{i}().}
\usage{
figure12.1()
figure12.2()
figure12.3()
figure12.4()
figure12.5()
figure12.6()
figure12.7()
figure3.1()
figure3.2()
figure3.3()
figure3.4()
figure3.5()
figure3.6()
figure3.7()
figure3.8()
figure3.9()
figure3.10()
figure3.11()
figure3.12()
figure3.13()
figure3.14()
figure3.15()
figure3.16()
figure3.17()
figure3.18()
figure3.19()
figure3.20()
figure3.21()
figure3.22()
figure3.23()
figure3.24()
figure3.25()
figure3.26()
figure3.27()
figure3.28()
figure3.29()
figure6.1()
figure6.2()
figure6.3()
figure6.4()
figure6.5()
figure6.6()
figure6.7()
figure6.8()
figure6.9()
figure6.10()
figure6.11()
figure6.12()
figure6.13()
figure6.14()
figure6.15()
figure6.16()
figure6.17()
figure6.18()
figure6.19()
figure6.20()
figure6.21()
figure6.22()
figure6.23()
figure6.24()
figure6.25()
figure6.26()
figure1.2()
figure1.3()
figure1.4()
figure1.5()
figure1.6()
figure1.7()
figure1.8()
figure1.9()
figure1.10()
figure1.11()
figure1.12()
figure1.13()
figure13.1()
figure13.2()
figure13.3()
figure13.4()
figure13.5()
figure13.6()
figure13.7()
figure13.8()
figure13.9()
figure13.10()
figure13.11()
figure13.12()
figure13.13()
figure13.14()
figure5.1()
figure5.2()
figure5.3()
figure5.4()
figure5.5()
figure5.6()
figure5.7()
figure5.8()
figure5.9()
figure5.10()
figure5.11()
figure5.12()
figure5.13()
figure5.14()
figure5.15()
figure5.16()
figure11.1()
figure11.2()
figure11.3()
figure11.4()
figure11.5()
figure11.6()
figure11.7()
figure11.8()
figure11.9()
figure11.10()
figure11.11()
figure11.12()
figure11.13()
figure11.14()
figure8.1()
figure8.3()
figure8.4()
figure8.5()
figure8.7()
figure8.8()
figure8.9()
figure8.11()
figure8.13()
figure8.14()
figure8.16()
figure8.17()
figure8.18()
figure8.20()
figure8.21()
figure8.22()
figure8.23()
figure8.25()
figure8.26()
figure8.28()
figure8.30()
figure8.32()
figure1.1()
figure7.1()
figure7.2()
figure7.3()
figure7.4()
figure7.5()
figure7.6()
figure7.7()
figure7.8()
figure7.9()
figure7.10()
figure7.11()
figure7.12()
figure7.13()
figure7.14()
figure7.15()
figure7.16()
figure1.14()
figure10.1()
figure10.2()
figure10.3()
figure10.4()
figure10.5()
figure10.6()
figure10.7()
figure10.8()
figure10.9()
figure10.10()
figure10.11()
figure10.12()
figure10.13()
figure2.1()
figure2.2()
figure2.3()
figure2.4()
figure2.5()
figure2.6()
figure2.7()
figure2.8()
figure2.9()
figure2.10()
figure4.1()
figure4.2()
figure4.3()
figure4.4()
figure4.5()
figure4.6()
figure4.7()
figure4.8()
figure4.9()
figure4.10()
figure4.11()
figure4.12()
}
\examples{
\donttest{
WD <- getwd()
setwd(tempdir())
grid.newpage(); figure12.1()
grid.newpage(); figure12.2()
grid.newpage(); figure12.3()
grid.newpage(); figure12.4()
grid.newpage(); figure12.5()
grid.newpage(); figure12.6()
grid.newpage(); figure12.7()
grid.newpage(); figure3.1()
grid.newpage(); figure3.2()
grid.newpage(); figure3.3()
grid.newpage(); figure3.4()
grid.newpage(); figure3.5()
grid.newpage(); figure3.6()
grid.newpage(); figure3.7()
grid.newpage(); figure3.8()
grid.newpage(); figure3.9()
grid.newpage(); figure3.10()
grid.newpage(); figure3.11()
grid.newpage(); figure3.12()
grid.newpage(); figure3.13()
grid.newpage(); figure3.14()
grid.newpage(); figure3.15()
grid.newpage(); figure3.16()
grid.newpage(); figure3.17()
grid.newpage(); figure3.18()
grid.newpage(); figure3.19()
grid.newpage(); figure3.20()
grid.newpage(); figure3.21()
grid.newpage(); figure3.22()
grid.newpage(); figure3.23()
grid.newpage(); figure3.24()
grid.newpage(); figure3.25()
grid.newpage(); figure3.26()
grid.newpage(); figure3.27()
grid.newpage(); figure3.28()
grid.newpage(); figure3.29()
grid.newpage(); figure6.1()
grid.newpage(); figure6.2()
grid.newpage(); figure6.3()
grid.newpage(); figure6.4()
grid.newpage(); figure6.5()
grid.newpage(); figure6.6()
grid.newpage(); figure6.7()
grid.newpage(); figure6.8()
grid.newpage(); figure6.9()
grid.newpage(); figure6.10()
grid.newpage(); figure6.11()
grid.newpage(); figure6.12()
grid.newpage(); figure6.13()
grid.newpage(); figure6.14()
grid.newpage(); figure6.15()
grid.newpage(); figure6.16()
grid.newpage(); figure6.17()
grid.newpage(); figure6.18()
grid.newpage(); figure6.19()
grid.newpage(); figure6.20()
grid.newpage(); figure6.21()
grid.newpage(); figure6.22()
grid.newpage(); figure6.23()
grid.newpage(); figure6.24()
grid.newpage(); figure6.25()
grid.newpage(); figure6.26()
grid.newpage(); figure1.2()
grid.newpage(); figure1.3()
grid.newpage(); figure1.4()
grid.newpage(); figure1.5()
grid.newpage(); figure1.6()
grid.newpage(); figure1.7()
grid.newpage(); figure1.8()
grid.newpage(); figure1.9()
grid.newpage(); figure1.10()
grid.newpage(); figure1.11()
grid.newpage(); figure1.12()
grid.newpage(); figure1.13()
grid.newpage(); figure13.1()
grid.newpage(); figure13.2()
grid.newpage(); figure13.3()
grid.newpage(); figure13.4()
grid.newpage(); figure13.5()
grid.newpage(); figure13.6()
grid.newpage(); figure13.7()
grid.newpage(); figure13.8()
grid.newpage(); figure13.9()
grid.newpage(); figure13.10()
grid.newpage(); figure13.11()
grid.newpage(); figure13.12()
grid.newpage(); figure13.13()
grid.newpage(); figure13.14()
grid.newpage(); figure5.1()
grid.newpage(); figure5.2()
grid.newpage(); figure5.3()
grid.newpage(); figure5.4()
grid.newpage(); figure5.5()
grid.newpage(); figure5.6()
grid.newpage(); figure5.7()
grid.newpage(); figure5.8()
grid.newpage(); figure5.9()
grid.newpage(); figure5.10()
grid.newpage(); figure5.11()
grid.newpage(); figure5.12()
grid.newpage(); figure5.13()
grid.newpage(); figure5.14()
grid.newpage(); figure5.15()
grid.newpage(); figure5.16()
grid.newpage(); figure11.1()
grid.newpage(); figure11.2()
grid.newpage(); figure11.3()
grid.newpage(); figure11.4()
grid.newpage(); figure11.5()
grid.newpage(); figure11.6()
grid.newpage(); figure11.7()
grid.newpage(); figure11.8()
grid.newpage(); figure11.9()
grid.newpage(); figure11.10()
grid.newpage(); figure11.11()
grid.newpage(); figure11.12()
grid.newpage(); figure11.13()
grid.newpage(); figure11.14()
grid.newpage(); figure8.1()
grid.newpage(); figure8.3()
grid.newpage(); figure8.4()
grid.newpage(); figure8.5()
grid.newpage(); figure8.7()
grid.newpage(); figure8.8()
grid.newpage(); figure8.9()
grid.newpage(); figure8.11()
grid.newpage(); figure8.13()
grid.newpage(); figure8.14()
grid.newpage(); figure8.16()
grid.newpage(); figure8.17()
grid.newpage(); figure8.18()
grid.newpage(); figure8.20()
grid.newpage(); figure8.21()
grid.newpage(); figure8.22()
grid.newpage(); figure8.23()
grid.newpage(); figure8.25()
grid.newpage(); figure8.26()
grid.newpage(); figure8.28()
grid.newpage(); figure8.30()
grid.newpage(); figure8.32()
grid.newpage(); figure1.1()
grid.newpage(); figure7.1()
grid.newpage(); figure7.2()
grid.newpage(); figure7.3()
grid.newpage(); figure7.4()
grid.newpage(); figure7.5()
grid.newpage(); figure7.6()
grid.newpage(); figure7.7()
grid.newpage(); figure7.8()
grid.newpage(); figure7.9()
grid.newpage(); figure7.10()
grid.newpage(); figure7.11()
grid.newpage(); figure7.12()
grid.newpage(); figure7.13()
grid.newpage(); figure7.14()
grid.newpage(); figure7.15()
grid.newpage(); figure7.16()
grid.newpage(); figure1.14()
grid.newpage(); figure10.1()
grid.newpage(); figure10.2()
grid.newpage(); figure10.3()
grid.newpage(); figure10.4()
grid.newpage(); figure10.5()
grid.newpage(); figure10.6()
grid.newpage(); figure10.7()
grid.newpage(); figure10.8()
grid.newpage(); figure10.9()
grid.newpage(); figure10.10()
grid.newpage(); figure10.11()
grid.newpage(); figure10.12()
grid.newpage(); figure10.13()
grid.newpage(); figure2.1()
grid.newpage(); figure2.2()
grid.newpage(); figure2.3()
grid.newpage(); figure2.4()
grid.newpage(); figure2.5()
grid.newpage(); figure2.6()
grid.newpage(); figure2.7()
grid.newpage(); figure2.8()
grid.newpage(); figure2.9()
grid.newpage(); figure2.10()
grid.newpage(); figure4.1()
grid.newpage(); figure4.2()
grid.newpage(); figure4.3()
grid.newpage(); figure4.4()
grid.newpage(); figure4.5()
grid.newpage(); figure4.6()
grid.newpage(); figure4.7()
grid.newpage(); figure4.8()
grid.newpage(); figure4.9()
grid.newpage(); figure4.10()
grid.newpage(); figure4.11()
grid.newpage(); figure4.12()
if (!is.null(WD)) setwd(WD)
}
}
\keyword{internal}
