\name{figures}
\alias{figureB.1}
\alias{figureB.2}
\alias{figure3.1}
\alias{figure3.2}
\alias{figure3.3}
\alias{figure3.4}
\alias{figure3.5}
\alias{figure3.6}
\alias{figure3.7}
\alias{figure3.8}
\alias{figure3.9}
\alias{figure3.10}
\alias{figure3.11}
\alias{figure3.12}
\alias{figure3.13}
\alias{figure3.14}
\alias{figure3.15}
\alias{figure3.16}
\alias{figure3.17}
\alias{figure3.18}
\alias{figure3.19}
\alias{figure3.20}
\alias{figure3.21}
\alias{figure3.22}
\alias{figure3.23}
\alias{figure3.24}
\alias{figure3.25}
\alias{figure3.26}
\alias{figure3.27}
\alias{figure3.28}
\alias{figure3.29}
\alias{figure3.30}
\alias{figure5.1}
\alias{figure5.2}
\alias{figure5.3}
\alias{figure5.4}
\alias{figure5.5}
\alias{figure5.6}
\alias{figure5.7}
\alias{figure5.8}
\alias{figure5.9}
\alias{figure5.10}
\alias{figure5.11}
\alias{figure5.12}
\alias{figure5.13}
\alias{figure5.14}
\alias{figure5.15}
\alias{figure5.16}
\alias{figure5.17}
\alias{figure5.18}
\alias{figure5.19}
\alias{figure5.20}
\alias{figure5.21}
\alias{figure5.22}
\alias{figure1.2}
\alias{figure1.3}
\alias{figure1.4}
\alias{figure1.5}
\alias{figure1.6}
\alias{figure1.7}
\alias{figure1.8}
\alias{figure1.9}
\alias{figure1.10}
\alias{figure1.11}
\alias{figure7.1}
\alias{figure7.3}
\alias{figure7.5}
\alias{figure7.6}
\alias{figure7.7}
\alias{figure7.13}
\alias{figure7.14}
\alias{figure7.18}
\alias{figure7.20}
\alias{figure7.22}
\alias{figure1.1}
\alias{figure6.1}
\alias{figure6.2}
\alias{figure6.3}
\alias{figure6.4}
\alias{figure6.5}
\alias{figure6.6}
\alias{figure6.7}
\alias{figure6.8}
\alias{figure6.9}
\alias{figure6.10}
\alias{figure1.12}
\alias{figure2.1}
\alias{figure2.2}
\alias{figure2.3}
\alias{figure2.4}
\alias{figure2.5}
\alias{figure2.6}
\alias{figure2.7}
\alias{figure2.8}
\alias{figure4.1}
\alias{figure4.2}
\alias{figure4.3}
\alias{figure4.4}
\alias{figure4.5}
\alias{figure4.6}
\alias{figure4.7}
\alias{figure4.8}
\alias{figure4.9}
\title{Figures for R Graphics}
\description{Functions to produce the figures from the book R Graphics.  Function \emph{i} is produced by figure\emph{i}().}
\synopsis{
figureB.1()
figureB.2()
figure3.1()
figure3.2()
figure3.3()
figure3.4()
figure3.5()
figure3.6()
figure3.7()
figure3.8()
figure3.9()
figure3.10()
figure3.11()
figure3.12()
figure3.13()
figure3.14()
figure3.15()
figure3.16()
figure3.17()
figure3.18()
figure3.19()
figure3.20()
figure3.21()
figure3.22()
figure3.23()
figure3.24()
figure3.25()
figure3.26()
figure3.27()
figure3.28()
figure3.29()
figure3.30()
figure5.1()
figure5.2()
figure5.3()
figure5.4()
figure5.5()
figure5.6()
figure5.7()
figure5.8()
figure5.9()
figure5.10()
figure5.11()
figure5.12()
figure5.13()
figure5.14()
figure5.15()
figure5.16()
figure5.17()
figure5.18()
figure5.19()
figure5.20()
figure5.21()
figure5.22()
figure1.2()
figure1.3()
figure1.4()
figure1.5()
figure1.6()
figure1.7()
figure1.8()
figure1.9()
figure1.10()
figure1.11()
figure7.1()
figure7.3()
figure7.5()
figure7.6()
figure7.7()
figure7.13()
figure7.14()
figure7.18()
figure7.20()
figure7.22()
figure1.1()
figure6.1()
figure6.2()
figure6.3()
figure6.4()
figure6.5()
figure6.6()
figure6.7()
figure6.8()
figure6.9()
figure6.10()
figure1.12()
figure2.1()
figure2.2()
figure2.3()
figure2.4()
figure2.5()
figure2.6()
figure2.7()
figure2.8()
figure4.1()
figure4.2()
figure4.3()
figure4.4()
figure4.5()
figure4.6()
figure4.7()
figure4.8()
figure4.9()
}
\usage{figure1.1()}
\keyword{internal}
