% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid.imageGrob.R
\name{imageGrob}
\alias{imageGrob}
\title{imageGrob}
\usage{
imageGrob(nrow, ncol, cols, byrow = TRUE, name = NULL, gp = NULL,
  vp = NULL)
}
\arguments{
\item{nrow}{Number of rows/columns in the image.}

\item{ncol}{Number of rows/columns in the image.}

\item{cols}{Colors for the cells of the image (will be recycled).}

\item{byrow}{Logical value indicating whether colors should be allocated to
cells across rows or down columns.}

\item{name}{A character name for the grob.}

\item{gp}{A \link[grid:gpar]{grid::gpar} object containing graphical parameter settings or NULL.}

\item{vp}{A \link[grid:viewport]{grid::viewport} or \code{NULL}.}
}
\description{
imageGrob
}
