% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid.ozFun.R
\name{ozGrob}
\alias{ozGrob}
\alias{grid.ozGrob}
\title{ozGrob}
\usage{
ozGrob(ozRegion, name = NULL, gp = NULL, vp = NULL)

grid.ozGrob(...)
}
\arguments{
\item{ozRegion}{An object created using the \code{ozRegion} function from the
\pkg{oz} package.}

\item{name}{A character name for the grob.}

\item{gp}{A \link[grid:gpar]{grid::gpar} object containing graphical parameter settings or NULL.}

\item{vp}{A \link[grid:viewport]{grid::viewport} or \code{NULL}.}

\item{...}{Arguments to be passed to \link{ozGrob}.}
}
\description{
ozGrob
}
