\name{splitString}
\alias{splitString}
\title{Split text into multiple lines}
\description{Splits a single string into multiple lines
(by inserting line breaks) so that the output will
fit within the current viewport.}
\usage{
splitString(text)
}
\arguments{
  \item{text}{The string to split.}
}
\keyword{dplot}
