\name{splitTextGrob}
\alias{splitTextGrob}
\title{Split text into multiple lines}
\description{Splits a single string into multiple lines
(by inserting line breaks) so that the output will
fit within the current viewport.}
\usage{
splitTextGrob(text, ...)
}
\arguments{
  \item{text}
       {The string to split.}
  \item{\ldots}
       {Arguments passed to the {\tt grob} function.}
}
\keyword{dplot}
